refactor: chordUtils.tsを複数の責務別ファイルに分割
516行の巨大なchordUtils.tsファイルを保守性向上のため、以下の6つの小さなファイルに分割：

- musicConstants.ts: 音楽関連の定数（キー、拍子、セクション名など）
- chordCreation.ts: コード譜作成関数（createEmptyChordChart等）
- chordValidation.ts: バリデーション関数（各種validation関数）
- chordParsing.ts: コード解析関数（extractChordRoot等）
- transpose.ts: 移調関連関数（transposeChordName等）
- chartMigration.ts: データ移行処理（migrateChartData）

全てのインポートを適切に更新し、テスト（417/417成功）、lint、ビルドが正常に動作することを確認。
