+++ src/utils/chartMigration.ts
@@ -1,0 +1,15 @@
+import type { ChordChart } from '../types';
+
+// 既存データの移行処理：拍子に応じてbeatsPerBarを修正
+export const migrateChartData = (chart: ChordChart): ChordChart => {
+  const beatsPerBar = chart.timeSignature ? parseInt(chart.timeSignature.split('/')[0]) : 4;
+  
+  return {
+    ...chart,
+    sections: chart.sections?.map(section => ({
+      ...section,
+      // beatsPerBarが未定義、または4拍以外の拍子で4拍になっている場合は修正
+      beatsPerBar: (!section.beatsPerBar || (beatsPerBar !== 4 && section.beatsPerBar === 4)) ? beatsPerBar : section.beatsPerBar
+    })) || []
+  };
+};
